Refactoring checkPrerequisite
