+++ src/commands/app/start-services.ts
@@ -3,13 +3,12 @@ import path from "path";
 
 import { flags } from "@oclif/command";
 import chalk from "chalk";
-import Listr from "listr";
+
 import emoji from "node-emoji";
 
 import { Kommand } from "../../common";
 import { execute } from "../../support/execute";
-
-const MIN_DOCO_VERSION = "2.0.0";
+import { checkPrerequisites } from "../../support/docker/checkPrerequisites";
 
 const kuzzleServicesFile = `
 version: '3'
@@ -48,7 +47,7 @@ export default class AppStartServices extends Kommand {
     const docoFilename = path.join(this.kourouDir, "kuzzle-services.yml");
 
     const successfullCheck = this.flags.check
-      ? await this.checkPrerequisites()
+      ? await checkPrerequisites(this)
       : true;
 
     if (this.flags.check && successfullCheck) {
@@ -89,57 +88,6 @@ export default class AppStartServices extends Kommand {
       );
 
       this.log(chalk.grey(`  docker compose -f ${docoFilename} up\n`));
-    }
-  }
-
-  public async checkPrerequisites(): Promise<boolean> {
-    this.log(chalk.grey("Checking prerequisites..."));
-
-    const checks: Listr = new Listr([
-      {
-        title: `docker compose exists and the version is at least ${MIN_DOCO_VERSION}`,
-        task: async () => {
-          try {
-            const docov = await execute("docker", "compose", "version");
-            const matches = docov.stdout.match(/[^0-9.]*([0-9.]*).*/);
-            if (matches === null) {
-              throw new Error(
-                "Unable to read the version of Docker Compose. This is weird."
-              );
-            }
-            const docoVersion = matches.length > 0 ? matches[1] : null;
-
-            if (docoVersion === null) {
-              throw new Error(
-                "Unable to read the version of Docker Compose. This is weird."
-              );
-            }
-            try {
-              if (docoVersion < MIN_DOCO_VERSION) {
-                throw new Error(
-                  `Your version of Docker Compose (${docoVersion}) is below the required version (${MIN_DOCO_VERSION}).`
-                );
-              }
-            } catch (error: any) {
-              throw new Error(error);
-            }
-          } catch (error: any) {
-            throw new Error(
-              "Docker Compose couldn't be found. Are you sure Docker and the Compose plugin are installed?"
-            );
-          }
-        },
-      },
-    ]);
-
-    try {
-      await checks.run();
-
-      return true;
-    } catch (error: any) {
-      this.logKo(error.message);
-
-      return false;
     }
   }
 }
